fixed python include for robot commander
