+++ sr_robot_commander/CMakeLists.txt
@@ -4,8 +4,15 @@ project(sr_robot_commander)
 ## Find catkin macros and libraries
 ## if COMPONENTS list like find_package(catkin REQUIRED COMPONENTS xyz)
 ## is used, also find other catkin packages
-find_package(catkin REQUIRED COMPONENTS moveit_core moveit_ros_planning moveit_ros_planning_interface pluginlib)
-
+find_package(catkin REQUIRED COMPONENTS
+    rospy
+    moveit_commander
+    pluginlib
+    moveit_core
+    moveit_ros_planning_interface
+    moveit_ros_perception
+    interactive_markers
+)
 ## System dependencies are found with CMake's conventions
 # find_package(Boost REQUIRED COMPONENTS system)
 
@@ -79,8 +86,12 @@ catkin_package(
 #  INCLUDE_DIRS include
 #  LIBRARIES sr_robot_commander
   CATKIN_DEPENDS
+    rospy
+    moveit_commander
+    pluginlib
     moveit_core
     moveit_ros_planning_interface
+    moveit_ros_perception
     interactive_markers
 #  DEPENDS system_lib
 )
@@ -139,11 +150,10 @@ catkin_package(
 # )
 
 ## Mark other files for installation (e.g. launch and bag files, etc.)
-# install(FILES
-#   # myfile1
-#   # myfile2
-#   DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION}
-# )
+install(DIRECTORY
+  src
+  DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION}
+)
 
 #############
 ## Testing ##
