+++ sr_robot_commander/src/sr_robot_commander/sr_robot_commander.py
@@ -1,4 +1,4 @@
-# #!/usr/bin/python
+# #!/usr/bin/env python
 #
 # # Copyright 2015 Shadow Robot Company Ltd.
 # #
@@ -16,47 +16,48 @@
 # # with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
 #
-# from moveit_commander import MoveGroupCommander
-# from moveit_msgs.msg import RobotTrajectory
-#
-#
-# class SrRobotCommander(object):
-#     """
-#     Base class for hand and arm commanders
-#     """
-#
-#     def __init__(self, name):
-#         """
-#         Initialize MoveGroupCommander object
-#         @param name - name of the MoveIt group
-#         """
-#         self._move_group_commander = MoveGroupCommander(name)
-#
-#     def move_to_joint_value_target(self, joint_states, wait_result=True):
-#         """
-#         Set target of the robot's links and moves to it
-#         @param joint_states - dictionary with joint name and value
-#         @param wait_result - should method wait for movement end or not
-#         """
-#         self._move_group_commander.set_joint_value_target(joint_states)
-#         self._move_group_commander.go(wait=wait_result)
-#
-#     def _run_joint_trajectory(self, joint_trajectory):
-#         """
-#         Moves robot through all joint states with specified timeouts
-#         @param joint_trajectory - JointTrajectory class object. Represents trajectory of the joints which would be
-#         executed.
-#         """
-#         plan = RobotTrajectory()
-#         plan.joint_trajectory = joint_trajectory
-#         self._move_group_commander.execute(plan)
-#
-#     def _move_to_position_target(self, xyz, end_effector_link="", wait_result=True):
-#         """
-#         Specify a target position for the end-effector and moves to it
-#         @param xyz - new position of end-effector
-#         @param end_effector_link - name of the end effector link
-#         @param wait_result - should method wait for movement end or not
-#         """
-#         self._move_group_commander.set_position_target(xyz, end_effector_link)
-#         self._move_group_commander.go(wait=wait_result)
+import rospy
+
+from moveit_commander import MoveGroupCommander
+from moveit_msgs.msg import RobotTrajectory
+
+class SrRobotCommander(object):
+    """
+    Base class for hand and arm commanders
+    """
+
+    def __init__(self, name):
+        """
+        Initialize MoveGroupCommander object
+        @param name - name of the MoveIt group
+        """
+        self._move_group_commander = MoveGroupCommander(name)
+
+    def move_to_joint_value_target(self, joint_states, wait_result=True):
+        """
+        Set target of the robot's links and moves to it
+        @param joint_states - dictionary with joint name and value
+        @param wait_result - should method wait for movement end or not
+        """
+        self._move_group_commander.set_joint_value_target(joint_states)
+        self._move_group_commander.go(wait=wait_result)
+
+    def _run_joint_trajectory(self, joint_trajectory):
+        """
+        Moves robot through all joint states with specified timeouts
+        @param joint_trajectory - JointTrajectory class object. Represents trajectory of the joints which would be
+        executed.
+        """
+        plan = RobotTrajectory()
+        plan.joint_trajectory = joint_trajectory
+        self._move_group_commander.execute(plan)
+
+    def _move_to_position_target(self, xyz, end_effector_link="", wait_result=True):
+        """
+        Specify a target position for the end-effector and moves to it
+        @param xyz - new position of end-effector
+        @param end_effector_link - name of the end effector link
+        @param wait_result - should method wait for movement end or not
+        """
+        self._move_group_commander.set_position_target(xyz, end_effector_link)
+        self._move_group_commander.go(wait=wait_result)
